+++ src/frontend/main/src/utilfunctions/createPopup.ts
@@ -1,0 +1,17 @@
+export function createPopup(title: string = 'Authentication', location: string) {
+  const width = 500;
+  const height = 630;
+  const settings = [
+    ['width', width],
+    ['height', height],
+    ['left', window.innerWidth / 2 - width / 2],
+    ['top', window.innerHeight / 2 - height / 2],
+  ]
+    .map((x) => x.join('='))
+    .join(',');
+
+  const popup = window.open(location, title, settings);
+  if (!popup) return;
+
+  return popup;
+}
