+++ src/frontend/main/src/utilfunctions/login.ts
@@ -1,28 +1,13 @@
 import environment from "../environment";
+import { createPopup } from "./createPopup";
 declare global {
   interface Window {
-    authComplete:any;
+    authComplete: any;
   }
 }
 
 // Code taken from https://github.com/mapbox/osmcha-frontend/blob/master/src/utils/create_popup.js
-export function createPopup(title: string = 'Authentication', location: string) {
-  const width = 500;
-  const height = 630;
-  const settings = [
-    ['width', width],
-    ['height', height],
-    ['left', window.innerWidth / 2 - width / 2],
-    ['top', window.innerHeight / 2 - height / 2],
-  ]
-    .map((x) => x.join('='))
-    .join(',');
 
-  const popup = window.open(location, title, settings);
-  if (!popup) return;
-
-  return popup;
-}
 
 export const createLoginWindow = (redirectTo) => {
   const popup = createPopup('OSM auth', '');
@@ -53,7 +38,7 @@ export const createLoginWindow = (redirectTo) => {
               const params = new URLSearchParams({
                 username: userRes.user_data.username,
                 osm_oauth_token: res.access_token.access_token,
-                id:userRes.user_data.id,
+                id: userRes.user_data.id,
                 picture: userRes.user_data.img_url,
                 redirect_to: redirectTo,
               }).toString();
